refactor: 카드 추가 화면 관련 파일명 및 클래스명 변경
