+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Androidpayments">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.cardcreate.CreateCardActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Androidpayments">
@@ -23,7 +23,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.cardcategories.CardCategoriesActivity"
+            android:name=".ui.cards.CardsActivity"
             android:exported="false"
             android:theme="@style/Theme.Androidpayments" />
     </application>
