--- app/src/main/java/woowacourse/payments/ui/cardcategories/CardCategoriesScreen.kt
@@ -1,9 +1,0 @@
-package woowacourse.payments.ui.cardcategories
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun CardCategoriesScreen(modifier: Modifier) {
-
-}
